backend/learnera_app/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
coverage==7.9.2
cryptography==44.0.0
cssselect2==0.7.0
django-redis
daphne==4.1.2
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.56.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
loguru==0.7.3
lxml==5.3.1
msgpack==1.1.0
oscrypto==1.3.0
packaging==25.0
pdfkit==1.0.0
pillow==11.0.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pyOpenSSL==25.0.0
pypdf==5.3.0
pyphen==0.17.2
pytest==8.4.1
pytest-django==4.11.1
python-bidi==0.6.6
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.2
stripe==11.5.0
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.8.2
win32_setctime==1.2.0
xhtml2pdf==0.2.16
zope.interface==7.2
zopfli==0.2.3.post1
gunicorn
uvicorn